prediction-service/requirements.txt
flask==3.1.1
boto3==1.40.9
flask-cors==6.0.1
python-dotenv==1.1.1
joblib==1.4.2
scikit-learn==1.5.2
numpy==2.0.2
pandas==2.2.3
torch==2.5.1
